XDIAD/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anomalib==1.1.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.2.1
cycler==0.12.1
diffusers==0.28.0
docstring_parser==0.16
easydict==1.13
filelock==3.13.1
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.23.2
idna==3.4
imageio==2.34.1
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.2
jsonargparse==4.29.0
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
lightning==2.2.5
lightning-utilities==0.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pip==22.3.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.5
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.28.1
rich==13.7.1
rich-argparse==1.4.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
setuptools==65.5.1
shapely==2.0.4
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==1.0.3
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchmetrics==1.4.0.post0
torchvision==0.16.0+cu121
tqdm==4.66.4
triton==2.1.0
typeshed_client==2.5.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.13
wheel==0.38.4
yarl==1.9.4
zipp==3.19.0
